# docs/openapi/c2c-market-api-v1.yaml
openapi: 3.1.0
info:
  title: C2CMarket API
  version: 0.1.0
  summary: Backend contract foundation and official price lead slice.
  description: |
    This contract covers the first executable backend slice. C2CMarket is a price intelligence
    and community matching platform. It does not process payments, escrow funds, store third-party
    credentials, store plaintext native passwords, auto-deliver keys/accounts, or proxy upstream API
    traffic. Native C2CMarket account passwords are stored only as salted hashes.
servers:
  - url: http://localhost:8080
security:
  - sessionCookie: []
paths:
...
  /api/v1/me/announcements/{id}/seen:
    post:
      operationId: markAnnouncementSeen
      tags: [Announcements]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmptyRequest"
      responses:
        "200":
          description: Announcement first-seen receipt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementReceipt"
  /api/v1/me/announcements/{id}/read:
    post:
      operationId: markAnnouncementRead
      tags: [Announcements]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmptyRequest"
      responses:
        "200":
          description: Announcement read receipt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementReceipt"
  /api/v1/me/announcements/{id}/dismiss:
    post:
      operationId: dismissAnnouncement
      tags: [Announcements]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmptyRequest"
      responses:
        "200":
          description: Announcement home-banner dismissal receipt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementReceipt"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: c2c_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
...
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    AnnouncementReceipt:
      type: object
      required: [announcementId, announcementVersion]
      properties:
        announcementId:
          type: string
          format: uuid
        announcementVersion:
          type: integer
        firstSeenAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
        dismissedAt:
          type: string
          format: date-time
...
    EmptyRequest:
      type: object
      additionalProperties: false
      properties: {}
...